+++ sugar/sugar.go
@@ -31,8 +31,7 @@ func MakeRecursive(ctx context.Context, db ydb.Connection, pathToCreate string) 
 		i += x
 		sub := pathToCreate[:i+1]
 		info, err := db.Scheme().DescribePath(ctx, sub)
-		var opErr *errors.OperationError
-		if errors.As(err, &opErr) && opErr.Reason == errors.StatusSchemeError {
+		if ydb.IsOperationErrorSchemeError(err) {
 			err = db.Scheme().MakeDirectory(ctx, sub)
 			if err != nil {
 				return errors.WithStackTrace(err)
@@ -72,8 +71,7 @@ func RemoveRecursive(ctx context.Context, db ydb.Connection, pathToRemove string
 			dir, err = db.Scheme().ListDirectory(ctx, p)
 			return errors.WithStackTrace(err)
 		}, retry.WithIdempotent())
-		var opErr *errors.OperationError
-		if errors.As(err, &opErr) && opErr.Reason == errors.StatusSchemeError {
+		if ydb.IsOperationErrorSchemeError(err) {
 			return nil
 		}
 		if err != nil {