+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Refactored internal operation and transport errors
+
 ## v3.16.8
 * Added `config.ExcludeGRPCCodesForPessimization()` opttion for exclude some grpc codes from pessimization rules 
 * Refactored pessimization node conditions